requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyjwt
bcrypt
pydantic
pytest
httpx
pydantic-settings